+++ CHANGE_LOG.md
@@ -7,3 +7,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Added
 - managed-credentials plug-in which provides a DSL for re-usable loaded credentials from multiple sources
+- dependency-constraints plug-in which provides a DSL for applying dependency version constraints to all configurations in bulk from an external file